Convert UI to Dalamud WindowSystem
